upgrade: scaffold widget used from baseviewmodel
